increase digits range
increased supported digits detection from 9 to 19 digits.

+++ digits_counters.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-	int num;
+	unsigned long long int num;
 	cin >> num;
 	int digits = 1;
 	while (num / 10)
